Fixes main

+++ packages/postgres-database/main.py
@@ -11,7 +11,7 @@ import docker
 import sqlalchemy as sa
 from alembic import __version__ as __alembic_version__
 from alembic.config import Config as AlembicConfig
-from alembic.util import CommandError
+#from alembic.util import CommandError
 
 from simcore_postgres_database.settings import DSN, db_config, sqlalchemy_url
 
@@ -24,18 +24,31 @@ DISCOVERED = ".discovered-ignore.json"
 default_ini = here / 'alembic.ini'
 migration_dir = here / 'migration'
 
+def safe(if_fails_return=False):
+    def decorate(func):
+        def safe_func(*args, **kargs):
+            try:
+                res = func(*args, **kargs)
+                return res
+            except RuntimeError as err:
+                log.info("%s failed:  %s", func.__name__, str(err))
+            except Exception: #pylint: disable=broad-except
+                log.info("%s failed unexpectedly", func.__name__, exc_info=True)
+            return if_fails_return
+        return safe_func
+    return decorate
+
 #@retry(wait=wait_fixed(0.1), stop=stop_after_delay(60))
+@safe()
 def ping(url):
     """checks whether database is responsive"""
-    try:
-        engine = sa.create_engine(url)
-        conn = engine.connect()
-        conn.close()
-    except Exception:
-        log.error("Something went south ...")
-        return False
+    engine = sa.create_engine(url)
+    conn = engine.connect()
+    conn.close()
     return True
 
+
+@safe(if_fails_return=None)
 def get_service_published_port(service_name: str) -> int:
     client = docker.from_env()
     services = [x for x in client.services.list() if service_name in x.name]
@@ -51,6 +64,7 @@ def get_service_published_port(service_name: str) -> int:
 
 
 
+
 @click.group()
 def cli():
     """ Wraps alembic CLI to simplify database migration
@@ -59,39 +73,47 @@ def cli():
 
 
 
-
 @cli.command()
 def discover():
-    """ Discovers active databases """
+    """ Discovers active databases
+
+        - dumps first match
+        -
+    """
     click.echo('Discovering database ...')
+
     try:
         # db_config comes from configuration variables
         cfg = deepcopy(db_config)
         url = DSN.format(**cfg)
+
+        click.echo("Trying %s ..." % url)
         if not ping(url):
-            service_name = cfg['host']
-            cfg['port'] = get_service_published_port(service_name)
+            cfg['port'] = port = get_service_published_port(cfg['host'])
             cfg['host'] = "127.0.0.1"
             url = DSN.format(**cfg)
 
-            if not ping(url):
+            click.echo("Trying %s ..." % url)
+            if port is None or not ping(url):
                 raise RuntimeError
 
-        click.echo(f"{url} responded")
+        dumps = json.dumps(cfg, sort_keys=True, indent=4)
+        click.echo(f"Responded. \nDiscovered config: {dumps}")
+
         with open(DISCOVERED, 'w') as fh:
-            json.dump(cfg, fh, sort_keys=True, indent=4)
+            fh.write(dumps)
 
     except RuntimeError:
-        os.remove(DISCOVERED)
-        click.echo("database not found")
+        if os.path.exists(DISCOVERED):
+            os.remove(DISCOVERED)
+        click.echo("Database not found")
 
 
 def get_config():
     with open(DISCOVERED) as fh:
         cfg = json.load(fh)
 
-    with open(default_ini) as fh:
-        config = AlembicConfig(fh)
+    config = AlembicConfig(default_ini)
 
     config.set_main_option('script_location', str(migration_dir))
     config.set_main_option('sqlalchemy.url', DSN.format(**cfg))
@@ -100,13 +122,13 @@ def get_config():
 
 @cli.command()
 @click.option('-m', 'message')
-def revision(message):
+def review(message):
     """Auto-generates a new revison
 
         alembic revision --autogenerate -m "first tables"
     """
     click.echo('Auto-generates revision based on changes ')
-    import pdb; pdb.set_trace()
+
     config = get_config()
     alembic.command.revision(config, message,
         autogenerate=True, sql=False,
@@ -118,12 +140,9 @@ def revision(message):
 @click.argument('revision', default='head')
 def upgrade(revision):
     """Upgrades target database to a given revision"""
-    click.echo('Upgrading database')
-
-    import pdb; pdb.set_trace()
+    click.echo(f'Upgrading database to {revision} ...')
     config = get_config()
     alembic.command.upgrade(config, revision, sql=False, tag=None)
-
 
 
 if __name__ == '__main__':
